+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/XxdyAmVerifications1.cs
@@ -5,10 +5,10 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
 {
     public partial class XxdyAmVerifications1
     {
-        public decimal VerifyId { get; set; }
+        public int VerifyId { get; set; }
         public string? AssetNumber { get; set; }
         public string? AssetType { get; set; }
-        public decimal? BranchId { get; set; }
+        public int? BranchId { get; set; }
         public string? PlateNumber { get; set; }
         public string? Vin { get; set; }
         public string? ModelYear { get; set; }
@@ -41,9 +41,9 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
         public string? CustomField3 { get; set; }
         public string? CustomField4 { get; set; }
         public string? CustomField5 { get; set; }
-        public decimal? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
         public DateTime? CreationDate { get; set; }
-        public decimal? LastUpdatedBy { get; set; }
+        public int? LastUpdatedBy { get; set; }
         public DateTime? LastUpdateDate { get; set; }
     }
 }
