+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/XxdyAppDocumentSet.cs
@@ -5,7 +5,7 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
 {
     public partial class XxdyAppDocumentSet
     {
-        public decimal DocumentSetId { get; set; }
+        public int DocumentSetId { get; set; }
         public string DocumentSetName { get; set; } = null!;
         public string SetType { get; set; } = null!;
         public string EnabledFlag { get; set; } = null!;
@@ -16,10 +16,10 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
         public string? CustomField3 { get; set; }
         public string? CustomField4 { get; set; }
         public string? CustomField5 { get; set; }
-        public decimal? OrgId { get; set; }
-        public decimal? CreatedBy { get; set; }
+        public int? OrgId { get; set; }
+        public int? CreatedBy { get; set; }
         public DateTime? CreationDate { get; set; }
-        public decimal? LastUpdatedBy { get; set; }
+        public int? LastUpdatedBy { get; set; }
         public DateTime? LastUpdateDate { get; set; }
     }
 }
