+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/XxdyCatItemSubstitute.cs
@@ -5,11 +5,11 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
 {
     public partial class XxdyCatItemSubstitute
     {
-        public decimal SubstituteId { get; set; }
-        public decimal OriginalItemId { get; set; }
+        public int SubstituteId { get; set; }
+        public int OriginalItemId { get; set; }
         public string? ItemCode { get; set; }
         public string? ItemDescription { get; set; }
-        public decimal SubstituteItemId { get; set; }
+        public int SubstituteItemId { get; set; }
         public string? ItemSubstitute { get; set; }
         public string? SubstituteDesc { get; set; }
         public string? ContextValue { get; set; }
@@ -18,9 +18,9 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
         public string? CustomField3 { get; set; }
         public string? CustomField4 { get; set; }
         public string? CustomField5 { get; set; }
-        public decimal? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
         public DateTime? CreationDate { get; set; }
-        public decimal? LastUpdatedBy { get; set; }
+        public int? LastUpdatedBy { get; set; }
         public DateTime? LastUpdateDate { get; set; }
     }
 }
