+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/XxdyCatTyre.cs
@@ -5,7 +5,7 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
 {
     public partial class XxdyCatTyre
     {
-        public decimal LineId { get; set; }
+        public int LineId { get; set; }
         public string? Eco { get; set; }
         public string? Tipo { get; set; }
         public string? Dot { get; set; }
@@ -23,9 +23,9 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
         public string? CustomField3 { get; set; }
         public string? CustomField4 { get; set; }
         public string? CustomField5 { get; set; }
-        public decimal? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
         public DateTime? CreationDate { get; set; }
-        public decimal? LastUpdatedBy { get; set; }
+        public int? LastUpdatedBy { get; set; }
         public DateTime? LastUpdateDate { get; set; }
     }
 }
