+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/XxdyCatZone.cs
@@ -5,7 +5,7 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
 {
     public partial class XxdyCatZone
     {
-        public decimal ZoneId { get; set; }
+        public int ZoneId { get; set; }
         public string? ItemCode { get; set; }
         public string? ZoneType { get; set; }
         public string? ZoneName { get; set; }
@@ -17,9 +17,9 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
         public string? CustomField3 { get; set; }
         public string? CustomField4 { get; set; }
         public string? CustomField5 { get; set; }
-        public decimal? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
         public DateTime? CreationDate { get; set; }
-        public decimal? LastUpdatedBy { get; set; }
+        public int? LastUpdatedBy { get; set; }
         public DateTime? LastUpdateDate { get; set; }
         public string? ZoneSubtype { get; set; }
     }
