+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/XxdyCatZoneItemsInt.cs
@@ -5,9 +5,9 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
 {
     public partial class XxdyCatZoneItemsInt
     {
-        public decimal ZoneId { get; set; }
+        public int ZoneId { get; set; }
         public string? Zona { get; set; }
-        public decimal? AssemblyItemId { get; set; }
+        public int? AssemblyItemId { get; set; }
         public string? ItemCode { get; set; }
         public string? ItemDescription { get; set; }
         public string? ActivityCode { get; set; }
@@ -21,9 +21,9 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
         public string? CustomField3 { get; set; }
         public string? CustomField4 { get; set; }
         public string? CustomField5 { get; set; }
-        public decimal? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
         public DateTime? CreationDate { get; set; }
-        public decimal? LastUpdatedBy { get; set; }
+        public int? LastUpdatedBy { get; set; }
         public DateTime? LastUpdateDate { get; set; }
     }
 }
