+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/XxdySysItem.cs
@@ -5,7 +5,7 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
 {
     public partial class XxdySysItem
     {
-        public decimal InventoryItemId { get; set; }
+        public int InventoryItemId { get; set; }
         public string? ItemCode { get; set; }
         public string? Description { get; set; }
         public string? PrimaryUomCode { get; set; }
